requirements.txt

Pillow==9.2.0
face_recognition==1.3.0